Character Movement Adjustments & Aiming Fix
Movement speed has been reduced for all movement modes. Aiming has been
fixed so the line trace is drawn from the character to where the centre
of the camera is pointing. Attacking has been implemented, both while
aiming as well as a temporary melee hitbox that appears whilst not
aiming.